config: Remove old config
With the automatic conversion on load time, mixer ill always use the new
config format. Even if the mix is initialized with the legacy config, it
will be converted right away. As such, there is no need to allow the
creation of legacy configs anymore nor to check if it is being used.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>