Introducing the Delta-Manifests
Currently, when updating from version X to Y, in order to apply a delta, swupd
needs to download the Manifest from version X to check which file the
from-hash in delta file is pointing to. Downloading that Manifest some times
is even larger than the delta pack, so reducing the download size would make
a significant improvement in update time.

The Delta-Manifest is a Manifest included in delta packs with all information
from Manifest X in order to apply deltas, i.e., all files that are going to be
used to apply deltas.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>