Revert "helpers: Reduce download timeout"
This reverts commit 9cf854d8d6e45dcb087c0bf4f61bfc633c16e43b.

The reduction to the timeout has caused failures for users on slow
networks, so this must be reverted to the default values.

The reduction to the timeout was added to improve user experience when
connection failed due to a proxy error. It would take mixer too long to
fail and the error message didn't indicate what the problem could be.

The error is now properly indicated, so even if it takes a while for the
error to occur the first time, the user should be able to identify the
problem on the spot.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>