Release v5.2.1
Bug fixes:

- Do not user container for `build image`.
The build image command cannot run in a container because there are no
loopback devices to mount in a container environment. This change
forces mixer to run in native mode when build image is called.

- Revert download timeout to default values.
The timeout for downloads was causing mixer to fail in slow networks
because it was too short. This fix reverts the value back to default.

- Properly handle upstream URL with subpath.
Mixer were unable to properly handle downloads when the upstream URL
included subpaths since the subpath was discarded. This fix changes the
internal algorithm to preserve the subpath.

- Enforce --Offline flag.
During mixer initialization, some commands relied on a separate check to
determine if an online check was required. This check ignored the value
of the --offline check, causing some scenarios where mixer would try to
connect even when the flag was set. This fix gives priority to the flag
to make sure it is always respected when set.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>