builder: Use same commenting rules for package file and bundles
Make package file parsing ignore '#' and anything after that until the
end of the line.  Also trim spaces.  This makes the comment syntax the
same as bundle files.

At this point, one could consider trying to merge both parsing
functions, but given how simple the rules for this one are, kept them
separate.  We can revisit this later.