swupd: enable backwards-compatible build update
This commit enables mixer to write different manifest formats based on
the configured format value. When introducing a breaking change to a
manifest it will be registered with the format it requires. Existing
template is registered to everything through format25 to remain
compatible with Clear Linux OS format. Everything above the current
format gets the new template, in this case with the minversion added to
the MoM.

This means that format bumps do not need two different versions of mixer
installed since it will always know how to build the correct format
based on the configured value.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>