swupd: do not update delete versions over minversions
Fixes #455
Minversions should not update deleted files to the minversion as this
may cause swupd to attempt to delete content twice. This exacerbates
issues where a directory (/usr/local/bin) was mistakenly deleted in a
version because it was mistakenly added earlier. Updating the delete to
the minversion then causes the directory to be deleted off the client
system again.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>