cmd: Search mixer.state in config get/set
When running `config get` and `config set` subcommands, also look for
the property in `mixer.state` file. Both subcommands will look for the
value in `mixer.state` first and if the search fails they will look for
the value in `builder.conf`. If it fails a second time, the error is
printed.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>