config: Add BaseConfig interface
The base config interface is intended as a generic interface that can
perform reflection operations on TOMl config structures. This allows a
single API to handle both config and state files.

The SaveConfig() function in config was renamed to Save() to keep it in
line with the naming used in state and is now part of the BaseConfig
interface.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>