swupd: remove os-core-update-index file creation
The file was no longer being used due to the actual bundle-info files
being published in the os-core-update-index bundle. Because this file is
very large and requires an update every version remove this file.

This commit also fixes a bug where the os-core-update-index bundle was
being created for every build, regardless of changes. This was not a
visible bug when the os-core-update-index file itself was being created
because it had changes every time. Now that that file does not change
per release we should not be writing the manifest every time.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>