swupd: add contentsize hack to format switch
The contentsize hack exists to work around a bug in swupd-client
v3.15.3 which was released during format25 in Clear Linux. Because mixer
is maintaining backwards compatibility for old formats we can no longer
remove this hack when a new format bump happens in the OS. Instead we
need to apply this arbitrary maximum whenever mixer is building for the
bad format.

Since the manifest_format file no longer only switches for manifest
templates it was necessary to rename the file and associated _test.go
file to format_switch.go and format_switch_test.go, respectively.

Add testing for this functionality.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>