cmd: Break root preRun into smaller parts
The root preRun() function accumulated a lot of responsibility over time
that are better located into sub command. This PR deconstructs that
function in order to have a simpler codepath.

Now, the prerun performs only 3 actions:
- Start profiling if needed.
- Check if it is the root command.
- Check if the command should run on a container or not.

The container logic has also been simplified. All commands are not
treated as native commands. During the pre-run, if a command has the
container marker, the current execution is aborted and the command is
re-executed inside a container. This way, the native flag only needs to
be checked once during execution.

Most of the logic in the root preRun() were exclusive to the `build`
subcommand, so they were moved to the build preRun().

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>