config: Load offline property from mixer.state
The Mix.OFFLINE field was added to mixer.state to be used as default for
the offline mode. During the root PreRun, the value is read if the
--offline flag was not provided by the user. It the flag is provided,
the user value has priority.

During `mixer init` the current value of the --offline flag is used to
overwrite the default value provided by MixState. This makes the flag
persistent if used during init.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>