Revert "Use go mod vendor to manage vendor directory"
This reverts commit c6777502c8cd5caef74cad7b747361876bcf3f79.

To make this automatically packageable through autospec we need to do
some more work. The travis environment also needs to be trimmed down to
speed up the build, and building the package should work without network
connection. In the current state, it will not satisfy both travis and
autospec, so reverting this until it can be properly transitioned.