Enable iterative manifests with 0 file count
When a bundle is updated to only modify which bundles are included, an
iterative manifest can be generated with 0 files. This change adds a
type field to the Manifest struct to determine when an iterative
manifest is used and allows the creation of iterative manifests with
a 0 file count.

Fixes #516

Signed-off-by: John Akre <john.w.akre@intel.com>