bat: Fix bundle count
The bundle validation count only needs to account for the bundles that
have been directly added to the mix. With the fragmentation of the
default bundles into smaller bundles that are used as includes, the test
must ignore the included bundles and just count the ones that were
directly added.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>