cmd: Improve error message for invalid upstream
The `--clear-version` flag in mixer init accepts a string as parameter
to allow the user to set it to `latest`, which will auto fetch the
upstream version. For all other cases, this flag must be provided an
integer number. This patch clarifies the valid values for this flag both
in the error message and in the help description.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>