Support "optional" (also-add) bundles in manifests
An optional bundle is similar to the required bundles
(included bundles) but have a couple of differences.

 - Optional bundles will be installed by default when the main
   bundle that includes them is installed.
 - Users will have the option to skip the installation of optional
   bundles by using the flag --skip-optional/-o with bundle-add.
 - Optional bundles can be removed from the system even if the
   main bundle that installed them is still in the system.

From the perspective of Mixer a bundle will be considered as
optional whenever it is marked as also-add in the bundle spec
file. E.g. also-add(lib-openssl). And it should be added as
"also-add: <optional bundle>" in the bundle's manifest.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>