swupd: Skip included files when creating file records
When creating file records for a manifest, skipping files from the
bundle info file list that exist in one of the manifest's nested
includes removes the need for manifest file subtraction. This improves
performance because it reduces the number of unnecessary file hash
calculations and removes the need for bundle manifest file subtraction.

Signed-off-by: John Akre <john.w.akre@intel.com>