Delete index manifest
The os-core-update-index manifest was unused and injected into the mix
by mixer. This change removes support for the index manifest in formats
larger than 28 and deletes the index manifest when a mix passes format
28.

Additionally, this change consolidates the index manifest implementation
into the swupd package so that it can more easily be enabled or disabled.

Signed-off-by: John Akre <john.w.akre@intel.com>