Release 5.9.0
This release includes an optimization in the way file records are
generated in manifests. Instead of adding all files from a bundle and
later removing the files that already exist in that bundle dependencies,
mixer now preemptively checks if a file is part of a dependency before
generating its record. This greatly reduces the number of generated
records and speed build time.

Included patches:
- swupd: Skip included files when creating file records
- MCA: Parse rpm output with \a character

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>