travis: Parallelize bat tests
This PR breaks bat tests into individual test runs within travis
infrastructure. This is a workaround the current test failures on travis
due to infrastructure limitations and allow us to reset a single test in
the case of a failure.

This patch also disables create-mix-bump-version-add-remove-bundles test
since it is unable to complete in travis infra.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>