builder: Make it optional to add os-core-update
Usually a mix should include os-core-update to provide updates, but there
are uses of a mix without os-core-update in it. If users are going to flash
those images, use them as a container or a chroot they may want not to include
swupd at all in the mix.

This is also particularly important for 3rd-party support because we will need
to create mixes with only 3rd-party content, so no swupd.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>