create_manifests: Process Manifest.full even if not changed
Swupd isn't processing Manifest.full if there's no change in any file.
This just don't happen in regular mixes because os-core always change.
But in minimal mixes where os-core, os-core-update and os-core-update-index
the user may try to build an update without any changes in any bundle. In
this case the update will be dummy, but at least all files should be generated.

And in this case if full is not processed mixer is crashing.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>