docker: Remove built-in Docker support
Built-in Docker support was originally added to enable auto-format bump.
At that time, mixer could only handle one format version at a time, so
we needed 2 different binaries to cross a format boundary: 1 for the +10
and one for the +20. The docker support would allow us to download an
image with the right version for that given format and simplify the
process.

Mixer now has support for multiple formats and a single binary can
handle the whole format bump process. Built in docker image is no longer
needed.

Note: For Docker usage due to compatibility issues with the host system,
mixer-ci Docker image should be used instead.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>