Stop building bundles if missing dependencies
Currently, if dependencies are not resolved for a bundle, mixer ignores
it and continues processing other bundles. The user is not notified and
results in undesired output.

This fix ensures that the processing is stopped if dependencies are not
resolved and generates an error message.
Empty bundles are considered valid and does not stop processing.

Fixes #663

Signed-off-by: Ashlesha Atrey <ashlesha.atrey@intel.com>