Remove os-packages file
Currently, in order to improve performance, dnf install is not used
to extract the packages. Hence the database used to create os-packages
file is not being populated and thus the os-packages file is empty.
Moreover the os-packages file is not required and used by anybody.

Remove functions related to creation of the os-packages file.

Signed-off-by: Ashlesha Atrey <ashlesha.atrey@intel.com>