swupd: Retain previously deleted file entries
When the previous manifest's list of files was longer than the new
manifest's list of files there was a chance for previously deleted file
entries to be unexpectedly dropped. This change prevents previously
deleted files from dropping when the previous manifest's file list is
longer than the new manifest's list.

Fixes #684

Signed-off-by: John Akre <john.w.akre@intel.com>