Limit the amount of certain modifiers in packs
Systems using certain modifiers are currently not used as frequently
so in order to reduce pack size and improve update times skip those
modifiers from being considered for packs.

Currently the SSE version is skipped when either AVX2 or AVX512
versions exist and APX is always skipped.

This change does not apply to 0 packs as it would break installer
offline support.

Signed-off-by: William Douglas <william.douglas@intel.com>