Improve manifest include subtraction
Instead of giving up on directory subtraction entirely, try to
subtract directories that are in an included manifest *and* do not
have a non-directory file somewhere in their path.

For example:

bundle1:
/dir/file

bundle2: includes bundle1
/dir

In this case /dir would be subtracted from bundle2.

Also add a special case for subtracting directories if they are in
os-core as this reduces a lot of unused content from a bundle's
manifest.

Signed-off-by: William Douglas <william.douglas@intel.com>